### 2025-06-13 #8405 @editorialbot
[PRE REVIEW]: Risk of Bias: Explainable, human-in-the-loop software for general risk-of-bias assessment
**Submitting author:** @rob-luke (<a href="http://orcid.org/0000-0002-4930-8351">Robert Luke</a>)
**Repository:** https://github.com/rob-luke/risk-of-bias
**Branch with paper.md** (empty if default branch): manuscript
**Version:** v0.13.0
