### 2025-06-14 #8405 @rob-luke
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
>

I will attach dois asap (next 48 hours). Thanks for pointing it out
