+++ paper.bib
@@ -4,6 +4,7 @@
   journal={Cochrane handbook for systematic reviews of interventions},
   pages={205–228},
   year={2019},
+  doi = {10.1002/9781119536604.ch8},
   publisher={Wiley Online Library}
 }
 @article{whiting2016robis,
@@ -13,6 +14,7 @@
   volume={69},
   pages={225–234},
   year={2016},
+  doi = {10.1016/j.jclinepi.2015.06.005},
   publisher={Elsevier}
 }
 @article{mcguinness2021risk,
@@ -23,6 +25,7 @@
   number={1},
   pages={55–61},
   year={2021},
+  doi = {10.1002/jrsm.1411},
   publisher={Wiley Online Library}
 }
 @article{sterne2019rob,
@@ -31,6 +34,7 @@
   journal={bmj},
   volume={366},
   year={2019},
+  doi = {10.1136/bmj.l4898},
   publisher={British Medical Journal Publishing Group}
 }
 @article{savovic2014evaluation,
@@ -40,6 +44,7 @@
   volume={3},
   pages={1–12},
   year={2014},
+  doi = {10.1186/2046-4053-3-37},
   publisher={Springer}
 }
 @article{marshall2016robotreviewer,
@@ -50,6 +55,7 @@
   number={1},
   pages={193–201},
   year={2016},
+  doi = {10.1093/jamia/ocv044},
   publisher={Oxford University Press}
 }
 @article{kellermeyer2018covidence,
@@ -59,6 +65,7 @@
   volume={106},
   number={4},
   pages={580},
+  doi = { 10.5195/jmla.2018.513},
   year={2018}
 }
 @article{sterne2016robins,
@@ -67,6 +74,7 @@
   journal={bmj},
   volume={355},
   year={2016},
+  doi = {10.1136/bmj.i4919},
   publisher={British Medical Journal Publishing Group}
 }
 @article{whiting2011quadas,
@@ -77,6 +85,7 @@
   number={8},
   pages={529–536},
   year={2011},
+  doi = {10.7326/0003-4819-155-8-201110180-00009},
   publisher={American College of Physicians}
 }
 @article{wolff2019probast,
@@ -87,6 +96,7 @@
   number={1},
   pages={51–58},
   year={2019},
+  doi = {10.7326/M18-1376},
   publisher={American College of Physicians}
 }
 @article{page2021prisma,
@@ -95,6 +105,7 @@
   journal={bmj},
   volume={372},
   year={2021},
+  doi = {10.1136/bmj.n71},
   publisher={British Medical Journal Publishing Group}
 }
 @article{higgins2024tool,
@@ -104,5 +115,6 @@
   volume={186},
   pages={108602},
   year={2024},
+  doi = {10.1016/j.envint.2024.108602},
   publisher={Elsevier}
 }
