### 2025-06-15 3eb0ba7
docs: add API key instructions
