### 2025-07-13 ab91f84
Removed manual AI judgement
