### 2025-07-14 1a74d0e
feat: add dynamic domain judgements
