+++ risk_of_bias/export.py
@@ -26,6 +26,8 @@ def export_framework_as_markdown(framework: Framework, path: Path) -> None:
 
     for domain in framework.domains:
         lines.append(f"\n## Domain {domain.index}: {domain.name}")
+        if domain.judgement:
+            lines.append(f"Judgement: **{domain.judgement}**\n")
 
         if not domain.questions:
             lines.append("No questions defined.")
@@ -71,6 +73,8 @@ def export_framework_as_html(framework: Framework, path: Path) -> None:
 
     for domain in framework.domains:
         children.append(h2[f"Domain {domain.index}: {domain.name}"])
+        if domain.judgement:
+            children.append(p["Judgement: ", strong[domain.judgement]])
 
         if not domain.questions:
             children.append(p["No questions defined."])
