### 2025-07-14 c977201
feat: add dynamic domain judgements (#90)
