+++ risk_of_bias/frameworks/rob2/domains/_domain_6_overall.py
@@ -1,6 +1,13 @@
 from risk_of_bias.types._domain_types import Domain
 from risk_of_bias.types._question_types import Question
 
+
+def _compute_judgement(domain: Domain) -> str | None:
+    if domain.questions[0].response:
+        return domain.questions[0].response.response
+    return None
+
+
 q6_r = Question(
     question="Risk-of-bias judgement",
     allowed_answers=["Low", "High", "Some Concerns"],
@@ -26,4 +33,5 @@ overall_domain = Domain(
     questions=[q6_r, q6_o],
     name="Overall",
     index=6,
+    judgement_function=_compute_judgement,
 )
