+++ risk_of_bias/summary.py
@@ -70,11 +70,9 @@ def summarise_frameworks(
     it creates a standardized view that facilitates pattern recognition and evidence
     synthesis across multiple studies.
 
-    The function specifically looks for "Risk-of-bias judgement" questions within each
-    domain, which represent the final assessment conclusions after considering all
-    signaling questions and evidence. This approach aligns with established risk-of-bias
-    assessment methodologies where detailed questioning leads to domain-level
-    judgements.
+    The function reads the :pyattr:`~risk_of_bias.types._domain_types.Domain.judgement`
+    property of each domain. This property dynamically computes the risk judgement
+    based on the domain's current question responses.
 
     Key applications include:
     - Creating summary tables for systematic review publications
@@ -86,7 +84,8 @@ def summarise_frameworks(
     ----------
     frameworks : list[Framework]
         Completed framework assessments to summarise. These should contain
-        domain-level "Risk-of-bias judgement" responses.
+        domain-level judgements accessible via the
+        :pyattr:`~risk_of_bias.types._domain_types.Domain.judgement` property.
 
     Returns
     -------
@@ -108,12 +107,7 @@ def summarise_frameworks(
         manuscript = fw.manuscript or ""
         domain_results: dict[str, str | None] = {}
         for domain in fw.domains:
-            judgement = None
-            for question in domain.questions:
-                if question.question == "Risk-of-bias judgement" and question.response:
-                    judgement = question.response.response
-                    break
-            domain_results[domain.name] = judgement
+            domain_results[domain.name] = domain.judgement
         summary[manuscript] = domain_results
     return summary
 
