### 2025-07-14 9b4bc94
feat: implement algorithms
