### 2025-07-14 4fba01b
feat: add computed framework judgement
