+++ docs/api.md
@@ -1,8 +1,8 @@
 The API is comprised of a hierarchical set of types that mirror the structure of risk-of-bias assessment frameworks like RoB2. This design enables systematic evaluation of research manuscripts through structured questioning and evidence-based responses.
 
-- **Frameworks** define the top-level assessment structure containing multiple evaluation domains. In the RoB2 context, a Framework represents the complete "Risk of Bias 2" tool for randomized trials, organizing five primary assessment domains and an overall judgement domain into a cohesive evaluation instrument.
+- **Frameworks** define the top-level assessment structure containing multiple evaluation domains. In the RoB2 context, a Framework represents the complete "Risk of Bias 2" tool for randomized trials. It organises five primary domains and exposes a `judgement` property that returns the overall score derived from those domains.
 
-- **Domains** represent specific categories of bias within a framework. Each domain focuses on a particular aspect of study methodology that could introduce bias. Domains expose a `judgement` property that computes the final risk rating from the answers provided. For RoB2, the five primary domains are: (1) bias from randomization process, (2) bias from deviations from intended interventions, (3) bias from missing outcome data, (4) bias in outcome measurement, and (5) bias in selection of reported results. An additional "Overall" domain records the final risk-of-bias judgement and predicted direction of bias.
+- **Domains** represent specific categories of bias within a framework. Each domain focuses on a particular aspect of study methodology that could introduce bias. Domains expose a `judgement` property that computes the final risk rating from the answers provided. For RoB2 the five primary domains are: (1) bias from randomization process, (2) bias from deviations from intended interventions, (3) bias from missing outcome data, (4) bias in outcome measurement and (5) bias in selection of reported results. The "Overall" domain captures only the predicted direction of bias.
 
 - **Questions** are the specific signaling questions within each domain that guide the assessment process. These questions are designed to systematically evaluate potential sources of bias, with predefined allowed answers (typically "Yes", "Probably Yes", "Probably No", "No", "No Information", "Not Applicable"). But if the user requests no predefined answers (using None), then a string answer will be provided. Questions can be marked as required and are indexed for systematic processing.
 
