### 2025-07-14 6836209
fix: improve code clarity
