+++ risk_of_bias/frameworks/rob2/domains/_domain_6_overall.py
@@ -1,12 +1,6 @@
 from risk_of_bias.types._domain_types import Domain
 from risk_of_bias.types._question_types import Question
 
-
-def _compute_judgement(domain: Domain) -> str | None:
-    # NEED TO IMPLEMENT THIS
-    return None
-
-
 q6_o = Question(
     question="What is the overall predicted direction of bias for this outcome?",
     allowed_answers=[
@@ -17,13 +11,12 @@ q6_o = Question(
         "Away from null",
         "Unpredictable",
     ],
-    index=6.2,
+    index=6.1,
     is_required=False,
 )
 
 overall_domain = Domain(
-    questions=[q6_r, q6_o],
+    questions=[q6_o],
     name="Overall",
     index=6,
-    judgement_function=_compute_judgement,
 )
