### 2025-07-14 6dffaa0
feat: add framework judgement export and tests
