+++ risk_of_bias/export.py
@@ -24,6 +24,9 @@ def export_framework_as_markdown(framework: Framework, path: Path) -> None:
     if framework.manuscript:
         lines.append(f"\n**Manuscript:** {framework.manuscript}")
 
+    if framework.judgement:
+        lines.append(f"\n**Overall Judgement:** {framework.judgement}")
+
     for domain in framework.domains:
         lines.append(f"\n## Domain {domain.index}: {domain.name}")
         if domain.judgement:
@@ -70,6 +73,9 @@ def export_framework_as_html(framework: Framework, path: Path) -> None:
     # Add manuscript name if available
     if framework.manuscript:
         children.append(p[strong["Manuscript: "], framework.manuscript])
+
+    if framework.judgement:
+        children.append(p[strong["Overall Judgement: "], framework.judgement])
 
     for domain in framework.domains:
         children.append(h2[f"Domain {domain.index}: {domain.name}"])
