+++ tests/test_summary.py
@@ -9,6 +9,8 @@ from risk_of_bias.summary import (
     print_summary,
     summarise_frameworks,
 )
+from risk_of_bias.types._domain_types import Domain
+from risk_of_bias.types._framework_types import Framework
 from risk_of_bias.types._response_types import ReasonedResponseWithEvidenceAndRawData
 
 
@@ -78,6 +80,23 @@ def test_export_summary_uses_overall_domain(tmp_path: Path) -> None:
             domain.judgement_function = lambda _d: "High"
         else:
             domain.judgement_function = lambda _d: "Low"
+
+    summary = summarise_frameworks([framework])
+    export_path = tmp_path / "summary.csv"
+    export_summary(summary, export_path)
+
+    rows = [line.split(",") for line in export_path.read_text().splitlines()]
+    header = rows[0]
+    data = rows[1]
+    assert header[-1] == "Overall"
+    assert data[-1] == "High"
+
+
+def test_export_summary_uses_framework_judgement(tmp_path: Path) -> None:
+    domain1 = Domain(index=1, name="D1", judgement_function=lambda d: "Low")
+    domain2 = Domain(index=2, name="D2", judgement_function=lambda d: "High")
+    framework = Framework(name="T", domains=[domain1, domain2])
+    framework.manuscript = "studyC.pdf"
 
     summary = summarise_frameworks([framework])
     export_path = tmp_path / "summary.csv"
