+++ risk_of_bias/frameworks/rob2/domains/_domain_3_missing.py
@@ -67,13 +67,6 @@ q3_4 = Question(
     is_required=True,
 )
 
-q3_r = Question(
-    question="Risk-of-bias judgement",
-    allowed_answers=["Low", "High", "Some Concerns"],
-    index=3.5,
-    is_required=True,
-)
-
 q3_o = Question(
     question=(
         "Optional Questions: What is the predicted direction of bias due to "
@@ -91,8 +84,49 @@ q3_o = Question(
     is_required=False,
 )
 
+
+def _compute_judgement(domain: Domain) -> str | None:
+    """
+    Compute the risk of bias judgement for the Domain.
+    """
+    q1 = (
+        domain.questions[0].response.response if domain.questions[0].response else None
+    )  # 3.1
+    q2 = (
+        domain.questions[1].response.response if domain.questions[1].response else None
+    )  # 3.2
+    q3 = (
+        domain.questions[2].response.response if domain.questions[2].response else None
+    )  # 3.3
+    q4 = (
+        domain.questions[3].response.response if domain.questions[3].response else None
+    )  # 3.4
+
+    # Return None if any required question is unanswered
+    if None in (q1, q2, q3, q4):
+        return None
+
+    YES = {"Yes", "Probably Yes"}
+    NO = {"No", "Probably No"}
+
+    if q1 in YES:
+        return "Low"
+    else:
+        if q2 in YES:
+            return "Low"
+        else:
+            if q3 in NO:
+                return "Low"
+            else:
+                if q4 in NO:
+                    return "Some concerns"
+                else:
+                    return "High"
+
+
 domain_3_missing = Domain(
-    questions=[q3_1, q3_2, q3_3, q3_4, q3_r, q3_o],
+    questions=[q3_1, q3_2, q3_3, q3_4, q3_o],
     name="Missing outcome data",
     index=3,
+    judgement_function=_compute_judgement,
 )
