+++ risk_of_bias/frameworks/rob2/domains/_domain_6_overall.py
@@ -1,13 +1,6 @@
 from risk_of_bias.types._domain_types import Domain
 from risk_of_bias.types._question_types import Question
 
-q6_r = Question(
-    question="Risk-of-bias judgement",
-    allowed_answers=["Low", "High", "Some Concerns"],
-    index=6.1,
-    is_required=True,
-)
-
 q6_o = Question(
     question="What is the overall predicted direction of bias for this outcome?",
     allowed_answers=[
@@ -18,12 +11,12 @@ q6_o = Question(
         "Away from null",
         "Unpredictable",
     ],
-    index=6.2,
+    index=6.1,
     is_required=False,
 )
 
 overall_domain = Domain(
-    questions=[q6_r, q6_o],
+    questions=[q6_o],
     name="Overall",
     index=6,
 )
