### 2025-10-24 #8405 @tellarin
> 👋 [@tellarin](https://github.com/tellarin) - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow [these guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html).

Sure. Happy to review.

Best,
Börje
