### 2025-12-23 #9317 @galessiorob
👋 @JayeetaP @RISbO001 Checking in again on ETAs for you to get started on this review, I know it's holiday season but would be great to have some signal from you - please let me know if you have questions about the process.

@tellarin thanks for getting started, would love an ETA from you as well.

Thanks!
