### 2025-12-28 #9317 @RISbO001
For Software Paper review point, I am not able to find
1. "A formal software paper in the repository or linked documentation.", Summary for Non-specialist Audience is provided in documentation but not in the form of a peer-reviewed paper. So, it might come under non-formal paper.
2. Statement of Need: No software paper present, so no section exists.
3. State of the Field: No comparison to other tools (e.g., ROBIN-I, ROBVIS) in docs; documentation focuses on this tool’s features only.
4. Quality of Writing: Documentation is clear and useful, but again, there isn’t a separate published paper to evaluate.
5. References: The web documentation includes a reference to the standard Cochrane RoB2 publication. However, this is not within a formal paper - just docs.
